+++ DBModels/JobRecieptMaster.cs
@@ -12,9 +12,9 @@ namespace Transactiondetails.DBModels
         public string ProcessingPointCode { get; set; }
         public string FinancialYearCode { get; set; }
         public string BranchCode { get; set; }
-        public Int16 UserCode { get; set; }
+        public int UserCode { get; set; }
         public DateTime EntryDate { get; set; }
-        public Int16 ModiUserCode { get; set; }
+        public int ModiUserCode { get; set; }
         public DateTime ModiDate { get; set; }
 
     }
