+++ DBModels/System3.cs
@@ -4,7 +4,7 @@ namespace Transactiondetails.DBModels
 {
     public class System3
     {
-        public Int16 Field1 { get; set; }
+        public int Field1 { get; set; }
         public string Field2 { get; set; }
         public string Field7 { get; set; }
