+++ ViewModels/UserData.cs
@@ -5,7 +5,7 @@ namespace Transactiondetails.ViewModels
     public class UserData
     {
         public string UserName { get; set; }
-        public Int16 UserId { get; set; }
+        public int UserId { get; set; }
 
         public string Branch { get; set; }
         public string FYear { get; set; }
